@@ -1,6 +1,6 @@
 {
-  "name": "envshare",
-  "version": "0.1.0",
+  "name": "himitsu",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -8,35 +8,38 @@
     "start": "next start",
     "lint": "next lint",
     "test": "jest --collect-coverage",
-    "fmt": "pnpm rome check . --apply-suggested && pnpm rome format . --write"
+    "fmt": "pnpm biome check . && pnpm biome format --write ."
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.13",
-    "@next/font": "13.1.1",
-    "@tailwindcss/forms": "^0.5.3",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
-    "@upstash/redis": "^1.19.1",
-    "@vercel/analytics": "0.1.7-beta.1",
-    "@vercel/og": "^0.0.27",
-    "base-x": "^4.0.0",
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
-    "next": "13.1.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "4.9.4",
-    "zod": "^3.20.2"
+    "@heroicons/react": "^2.1.5",
+    "@next/font": "14.2.5",
+    "@tailwindcss/forms": "^0.5.7",
+    "@types/node": "22.0.0",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "@upstash/redis": "^1.33.0",
+    "@vercel/analytics": "1.3.1",
+    "@vercel/og": "^0.6.2",
+    "base-x": "^5.0.0",
+    "clsx": "^2.1.1",
+    "geist": "^1.3.1",
+    "next": "14.2.5",
+    "next-themes": "^0.3.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "tailwind-merge": "^2.4.0",
+    "typescript": "5.5.4",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@jest/globals": "^29.3.1",
-    "@types/jest": "^29.2.5",
-    "autoprefixer": "^10.4.13",
-    "jest": "^29.3.1",
-    "postcss": "^8.4.21",
-    "rome": "^11.0.0",
-    "tailwindcss": "^3.2.4",
-    "ts-jest": "^29.0.5"
+    "@biomejs/biome": "^1.8.3",
+    "@jest/globals": "^29.7.0",
+    "@types/jest": "^29.5.12",
+    "autoprefixer": "^10.4.19",
+    "jest": "^29.7.0",
+    "postcss": "^8.4.40",
+    "rome": "^12.1.3",
+    "tailwindcss": "^3.4.7",
+    "ts-jest": "^29.2.3"
   }
 }
