@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   experimental: {
-    appDir: true,
+    instrumentationHook: true,
   },
   images: {
     domains: ["twitter.com", "pbs.twimg.com"],
