@@ -1,4 +1,4 @@
-import { fromBase58, toBase58 } from "../util/base58";
+import { fromBase58, toBase58 } from "../lib/base58";
 import { ID_LENGTH, ENCRYPTION_KEY_LENGTH } from "./constants";
 /**
  * To share links easily, we encode the id, where the data is stored in redis, together with the secret encryption key.
