@@ -1,4 +1,4 @@
-import { toBase58 } from "../util/base58";
+import { toBase58 } from "../lib/base58";
 import { ID_LENGTH } from "./constants";
 
 export function generateId(): string {
